node typo done 🧶

+++ 6-TypeNode/backend/services/image-service.ts
@@ -1,26 +1,29 @@
-const fs = require('fs').promises
+import fs from 'fs/promises'
 
-const saveImage = async (file) => {
-  let tmp_path = file.path;
-  let target_path = 'public/images/' + file.originalname;
-  const data = await fs.readFile(tmp_path)
-  await fs.writeFile(target_path, data)
-  await fs.unlink(tmp_path)
-  return true
+const saveImage = async (file: Express.Multer.File | undefined) => {
+  if(file){
+    let tmp_path = file.path;
+    let target_path = 'public/images/' + file.originalname;
+    const data = await fs.readFile(tmp_path)
+    await fs.writeFile(target_path, data)
+    await fs.unlink(tmp_path)
+    return true
+  }
+  return false
 }
 
-const deleteImage = async(fileName) => {
+const deleteImage = async(fileName: string) => {
   let path = 'public/images/'+ fileName
   await fs.unlink(path)
   return true
 }
 
-const getImage = async(fileName) => {
+const getImage = async(fileName: string) => {
   let path = 'public/images/'+ fileName
   return await fs.readFile(path)
 }
 
-module.exports = {
+export default {
   saveImage,
   deleteImage,
   getImage
